+++ Assets/Tests/Editor/PredictorTest.cs
@@ -15,7 +15,7 @@ namespace Function.Tests {
         private Function fxn;
 
         [SetUp]
-        public void Before () => fxn = FunctionUnity.Create(url: @"https://api.fxn.dev/graph");
+        public void Before () => fxn = FunctionUnity.Create(url: @"https://api.fxn.dev");
 
         [Test(Description = @"Should retrieve a valid predictor")]
         public async Task RetrievePredictor () {