+++ Packages/ai.fxn.fxn3d/Runtime/Internal/Function.cs
@@ -111,6 +111,19 @@ namespace Function.Internal {
         );
         [DllImport(Assembly, EntryPoint = @"FXNValueCreateNull")]
         public static extern Status CreateNullValue (out IntPtr value);
+        [DllImport(Assembly, EntryPoint = @"FXNValueCreateBySerializingValue")]
+        public static extern Status CreateSerializedValue (
+            this IntPtr value,
+            ValueFlags flags,
+            out IntPtr result
+        );
+        [DllImport(Assembly, EntryPoint = @"FXNValueCreateByDeserializingValue")]
+        public static extern Status CreateDeserializedValue (
+            this IntPtr value,
+            Dtype type,
+            ValueFlags flags,
+            out IntPtr result
+        );
         #endregion