+++ Packages/ai.fxn.fxn3d/Runtime/Types/Image.cs
@@ -40,7 +40,7 @@ namespace Function.Types {
         /// <summary>
         /// Create an image.
         /// </summary>
-        /// <param name="data">Image pixel buffer.</param>
+        /// <param name="data">Pixel buffer. The pixel buffer format MUST be `R8`, `RGB888`, or `RGBA8888`.</param>
         /// <param name="width">Image width.</param>
         /// <param name="height">Image height.</param>
         /// <param name="channels">Image channels.</param>
@@ -51,19 +51,27 @@ namespace Function.Types {
             this.height = height;
             this.channels = channels;
         }
-        #endregion
 
-
-        #region --Operations--
-        private readonly byte* nativeData;
-
-        public unsafe Image (byte* data, int width, int height, int channels) { // Zero copy into `FXNValue`
+        /// <summary>
+        /// Create an image from a pixel buffer.
+        /// NOTE: DO NOT use this overload unless you absolutely know what you are doing.
+        /// </summary>
+        /// <param name="data">Pixel buffer. The pixel buffer format MUST be `R8`, `RGB888`, or `RGBA8888`.</param>
+        /// <param name="width">Image width.</param>
+        /// <param name="height">Image height</param>
+        /// <param name="channels">Image channels.</param>
+        public unsafe Image (byte* data, int width, int height, int channels) { // Enables zero copy into `FXNValue`
             this.data = null!;
             this.nativeData = data;
             this.width = width;
             this.height = height;
             this.channels = channels;
         }
+        #endregion
+
+
+        #region --Operations--
+        private readonly byte* nativeData;
 
         public ref byte GetPinnableReference () => ref (nativeData == null ? ref data[0] : ref *nativeData);
         #endregion